[not verified] Move controls to separate component
