+++ projects/plugins/jetpack/extensions/blocks/mailchimp/edit.js
@@ -4,15 +4,7 @@
 import apiFetch from '@wordpress/api-fetch';
 import classnames from 'classnames';
 import { __ } from '@wordpress/i18n';
-import {
-	Button,
-	ExternalLink,
-	PanelBody,
-	Placeholder,
-	Spinner,
-	TextControl,
-	withNotices,
-} from '@wordpress/components';
+import { Button, Placeholder, Spinner, TextControl, withNotices } from '@wordpress/components';
 import { InnerBlocks, InspectorControls, RichText } from '@wordpress/block-editor';
 import { Fragment, Component } from '@wordpress/element';
 
@@ -20,7 +12,7 @@ import { Fragment, Component } from '@wordpress/element';
  * Internal dependencies
  */
 import { icon, innerButtonBlock } from '.';
-import MailchimpGroups from './mailchimp-groups';
+import { MailChimpBlockControls } from './controls';
 
 const API_STATE_LOADING = 0;
 const API_STATE_CONNECTED = 1;
@@ -170,69 +162,21 @@ class MailchimpSubscribeEdit extends Component {
 		);
 		const inspectorControls = (
 			<InspectorControls>
-				<PanelBody title={ __( 'Text Elements', 'jetpack' ) }>
-					<TextControl
-						label={ __( 'Email Placeholder', 'jetpack' ) }
-						value={ emailPlaceholder }
-						onChange={ this.updateEmailPlaceholder }
-					/>
-				</PanelBody>
-				<PanelBody title={ __( 'Notifications', 'jetpack' ) }>
-					<TextControl
-						label={ __( 'Processing text', 'jetpack' ) }
-						value={ processingLabel }
-						onChange={ this.updateProcessingText }
-					/>
-					<TextControl
-						label={ __( 'Success text', 'jetpack' ) }
-						value={ successLabel }
-						onChange={ this.updateSuccessText }
-					/>
-					<TextControl
-						label={ __( 'Error text', 'jetpack' ) }
-						value={ errorLabel }
-						onChange={ this.updateErrorText }
-					/>
-				</PanelBody>
-				<PanelBody title={ __( 'Mailchimp Groups', 'jetpack' ) }>
-					<MailchimpGroups
-						interests={ interests }
-						onChange={ ( id, checked ) => {
-							// Create a Set to insure no duplicate interests
-							const deDupedInterests = [ ...new Set( [ ...interests, id ] ) ];
-							// Filter the clicked interest based on checkbox's state.
-							const updatedInterests = deDupedInterests.filter( item =>
-								item === id && ! checked ? false : item
-							);
-							setAttributes( {
-								interests: updatedInterests,
-							} );
-						} }
-					/>
-					<ExternalLink href="https://mailchimp.com/help/send-groups-audience/">
-						{ __( 'Learn about groups', 'jetpack' ) }
-					</ExternalLink>
-				</PanelBody>
-				<PanelBody title={ __( 'Signup Location Tracking', 'jetpack' ) }>
-					<TextControl
-						label={ __( 'Signup Field Tag', 'jetpack' ) }
-						placeholder={ __( 'SIGNUP', 'jetpack' ) }
-						value={ signupFieldTag }
-						onChange={ value => setAttributes( { signupFieldTag: value } ) }
-					/>
-					<TextControl
-						label={ __( 'Signup Field Value', 'jetpack' ) }
-						placeholder={ __( 'website', 'jetpack' ) }
-						value={ signupFieldValue }
-						onChange={ value => setAttributes( { signupFieldValue: value } ) }
-					/>
-					<ExternalLink href="https://mailchimp.com/help/determine-webpage-signup-location/">
-						{ __( 'Learn about signup location tracking', 'jetpack' ) }
-					</ExternalLink>
-				</PanelBody>
-				<PanelBody title={ __( 'Mailchimp Connection', 'jetpack' ) }>
-					<ExternalLink href={ connectURL }>{ __( 'Manage Connection', 'jetpack' ) }</ExternalLink>
-				</PanelBody>
+				<MailChimpBlockControls
+					emailPlaceholder={ emailPlaceholder }
+					updateEmailPlaceholder={ this.updateEmailPlaceholder }
+					processingLabel={ processingLabel }
+					updateProcessingText={ this.updateProcessingText }
+					successLabel={ successLabel }
+					updateSuccessText={ this.updateSuccessText }
+					errorLabel={ errorLabel }
+					updateErrorText={ this.updateErrorText }
+					interests={ interests }
+					setAttributes={ this.props.setAttributes }
+					signupFieldTag={ signupFieldTag }
+					signupFieldValue={ signupFieldValue }
+					connectURL={ connectURL }
+				/>
 			</InspectorControls>
 		);
 		const blockClasses = classnames( className, {
